functions/requirements-dev.txt
firebase_functions~=0.1.0
google-cloud-firestore
python-dotenv

